backend/requirements.txt
pymupdf>=1.23.0
pdfplumber>=0.10.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
jinja2>=3.1.3
google-genai>=1.0.0
pydantic>=2.6.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
xhtml2pdf>=0.2.17
